Moves parser to separate module.
